Fixed assignment of predicates, and TPTP solving.
Predicates (well, terms of type prop), do not need to be assigned, or
rather their assignment need not be a point of choice, since it already
is a point of choice for the sat solver, so all we need to do is add the
corresponding assignment when the sat solver decides on it.

TPTP and dimacs format do not have explicit statements, thus we need to
add an explicite solve instructions in these cases.